requirements.txt
python-telegram-bot==22.5
SQLAlchemy==2.0.45
pydantic==2.12.5
pydantic-settings==2.12.0
apscheduler==3.10.4
python-dotenv==1.0.0
cryptography==46.0.3
librouteros==3.4.1

